style: refactor relevant components
